rename: 폴더명 대문자로 최종 수정
